consolidate theme generation code
