+++ sites/docs/src/lib/components/colors/colors.ts
@@ -1,6 +1,8 @@
 import { z } from "zod";
 
-import { colors } from "$lib/registry/colors.js";
+import { colorMapping, colors, type BaseColor } from "$lib/registry/colors.js";
+import template from "lodash.template";
+import { BASE_STYLES, BASE_STYLES_WITH_VARIABLES } from "$lib/registry/templates.js";
 
 const colorSchema = z.object({
 	name: z.string(),
@@ -84,4 +86,72 @@ function getForegroundFromBackground(rgb: string) {
 	const luminance = 0.2126 * toLinear(r) + 0.7152 * toLinear(g) + 0.0722 * toLinear(b);
 
 	return luminance > 0.179 ? "#000" : "#fff";
+}
+
+export function generateBaseColorTemplate(baseColor: BaseColor) {
+	const colorsData = getColorsData();
+
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	const base: Record<string, any> = {
+		inlineColors: {},
+		cssVars: {},
+	};
+	for (const [mode, values] of Object.entries(colorMapping)) {
+		base.inlineColors[mode] = {};
+		base.cssVars[mode] = {};
+		for (const [key, value] of Object.entries(values)) {
+			if (typeof value === "string") {
+				const resolvedColor = value.replace(/\{\{base\}\}-/g, `${baseColor}-`);
+				base.inlineColors[mode][key] = resolvedColor;
+
+				const [resolvedBase, scale] = resolvedColor.split("-");
+				const color = scale
+					? colorsData[resolvedBase].find(
+							// eslint-disable-next-line @typescript-eslint/no-explicit-any
+							(item: any) => item.scale === Number.parseInt(scale)
+						)
+					: colorsData[resolvedBase];
+				if (color) {
+					base.cssVars[mode][key] = color.oklch;
+				}
+			}
+		}
+	}
+
+	// Build css vars.
+	base.inlineColorsTemplate = template(BASE_STYLES)({});
+	base.cssVarsTemplate = template(BASE_STYLES_WITH_VARIABLES)({
+		colors: base.cssVars,
+	});
+
+	return base;
+}
+
+export function getColorsData() {
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	const colorsData: Record<string, any> = {};
+	for (const [color, value] of Object.entries(colors)) {
+		if (typeof value === "string") {
+			colorsData[color] = value;
+			continue;
+		}
+
+		if (Array.isArray(value)) {
+			colorsData[color] = value.map((item) => ({
+				...item,
+				oklch: item.oklch,
+			}));
+			continue;
+		}
+
+		if (typeof value === "object") {
+			colorsData[color] = {
+				...value,
+				oklch: value.oklch,
+			};
+			continue;
+		}
+	}
+
+	return colorsData;
 }
