+++ sites/docs/src/lib/registry/colors.ts
@@ -5,7 +5,7 @@ export type CompleteColorFormat = {
 	oklch: string;
 };
 
-type ColorScale = {
+export type ColorScale = {
 	scale: number;
 	hex: string;
 	rgb: string;
@@ -13,7 +13,7 @@ type ColorScale = {
 	oklch: string;
 };
 
-type Colors = {
+export type Colors = {
 	inherit: string;
 	current: string;
 	transparent: string;
@@ -42,6 +42,30 @@ type Colors = {
 	pink: ColorScale[];
 	rose: ColorScale[];
 };
+
+export type BaseColor =
+  | "slate"
+  | "gray"
+  | "zinc"
+  | "neutral"
+  | "stone"
+  | "red"
+  | "orange"
+  | "amber"
+  | "yellow"
+  | "lime"
+  | "green"
+  | "emerald"
+  | "teal"
+  | "cyan"
+  | "sky"
+  | "blue"
+  | "indigo"
+  | "violet"
+  | "purple"
+  | "fuchsia"
+  | "pink"
+  | "rose";
 
 export const colors: Colors = {
 	inherit: "inherit",
