+++ sites/docs/src/lib/registry/templates.ts
